Updates and comments

+++ src/image/Vector.java
@@ -1,9 +1,9 @@
 package image;
 
 public class Vector {
-	public float x, y, z;
+	public double x, y, z;
 	
-	public Vector(float x, float y){this.x = x; this.y = y;}
-	public Vector(float x, float y, float z){this.x = x; this.y = y; this.z = z;}	
+	public Vector(double x, double y){this.x = x; this.y = y;}
+	public Vector(double x, double y, double z){this.x = x; this.y = y; this.z = z;}	
 	
 }
